% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprk_missile_tests_df.R
\docType{data}
\name{dprk_missile_tests_df}
\alias{dprk_missile_tests_df}
\title{The CNS North Korea Missile Test Database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 18 columns.}
\source{
\url{https://www.nti.org/analysis/articles/cns-north-korea-missile-test-database/}
}
\usage{
dprk_missile_tests_df
}
\description{
The James Martin Center for Nonproliferation Studies (CNS) North Korea Missile Test Database is the first database to record flight tests of all missiles launched by North Korea capable of delivering a payload of at least 500 kilograms (1102.31 pounds) a distance of at least 300 kilometers (186.4 miles). The database captures advancements in North Korea's missile program by documenting all such tests since the first one occurred in April 1984, and will be routinely updated as events warrant.
}
\examples{
dprk_missile_tests_df

dplyr::glimpse(dprk_missile_tests_df)

}
\references{
\itemize{
\item Created by The James Martin Center for Nonproliferation Studies for the Nuclear Threat Initiative
}
}
\keyword{datasets}
