% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca_df.R
\docType{data}
\name{nca_df}
\alias{nca_df}
\title{Nuclear Cooperation Agreement (NCA) Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3152 rows and 7 columns.}
\source{
\url{http://www.matthewfuhrmann.com/datasets.html}
}
\usage{
nca_df
}
\description{
The NCA dataset contains information on nuclear cooperation agreements signed by countries from 1945 to 2003.  It includes more than 1,500 treaties in which states pledged to transfer nuclear technology, materials, or know-how. The dataset is based on the 2003 version of James Keeley's list of nuclear cooperation agreements.
}
\examples{
nca_df
dplyr::glimpse(nca_df)

}
\references{
\itemize{
\item Fuhrmann, Matthew. 2012. Atomic Assistance: How "Atoms for Peace" Programs Cause Nuclear Insecurity. Ithaca, NY: Cornell University Press.
}
}
\keyword{datasets}
