% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nl_1.2_df.R
\docType{data}
\name{nl_1.2_df}
\alias{nl_1.2_df}
\title{Nuclear Latency (NL) Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 253 rows and 20 columns.}
\source{
\url{http://www.matthewfuhrmann.com/datasets.html}
}
\usage{
nl_1.2_df
}
\description{
The NL dataset provides information on the global diffusion of latent nuclear capabilities from 1939 to 2012. It identifies all known uranium enrichment and plutonium reprocessing facilities that have been built globally during the nuclear age.
}
\examples{
nl_1.2_df
dplyr::glimpse(nl_1.2_df)

}
\references{
\itemize{
\item Fuhrmann, Matthew and Benjamin Tkach. 2015. "Almost Nuclear: Introducing the Nuclear Latency Dataset." Conflict Management and Peace Science 32 (4): 443-461.
}
}
\keyword{datasets}
