% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp_19.1_df.R
\docType{data}
\name{ucdp_19.1_df}
\alias{ucdp_19.1_df}
\title{UCDP Georeferenced Event Dataset (GED) Global version 19.1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 152616 rows and 42 columns.}
\source{
\url{https://ucdp.uu.se/downloads/index.html#ged_global/}
}
\usage{
ucdp_19.1_df
}
\description{
This dataset is UCDP's most disaggregated dataset, covering individual events of organized violence (phenomena of lethal violence occurring at a given time and place). These events are sufficiently fine-grained to be geo-coded down to the level of individual villages, with temporal durations disaggregated to single, individual days.
}
\examples{
ucdp_19.1_df

dplyr::glimpse(ucdp_19.1_df)

}
\references{
\itemize{
\item Sundberg, Ralph, and Erik Melander, 2013, “Introducing the UCDP Georeferenced Event Dataset”, Journal of Peace Research, vol.50, no.4, 523-532
\item Högbladh Stina, 2019, “UCDP GED Codebook version 19.1”, Department of Peace and Conflict Research, Uppsala University
}
}
\keyword{datasets}
