% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp_syria_df.R
\docType{data}
\name{ucdp_syria_df}
\alias{ucdp_syria_df}
\title{UCDP Georeferenced Event Dataset (GED) Syria 20160101-20191130 (652.1601.1911)}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16566 rows and 41 columns.}
\source{
\url{https://ucdp.uu.se/downloads/index.html#ged_global/}
}
\usage{
ucdp_syria_df
}
\description{
This dataset contains georeferenced events data for Syria, covering 20160101-20191130. The dataset includes the same variables as the UCDP GED Global 19.1 dataset, and the same Codebook applies to both datasets.
In contrast to UCDP GED Global 19.1, this dataset includes all clear events, meaning that dyads do not have to pass the threshold of 25 battle-related deaths during one calendar year, to be included (see the UCDP GED Codebook for definitions and more information on events, dyads, and the 25 threshold).
}
\examples{
ucdp_syria_df
dplyr::glimpse(ucdp_syria_df)

}
\references{
\itemize{
\item Sundberg, Ralph, and Erik Melander, 2013, “Introducing the UCDP Georeferenced Event Dataset”, Journal of Peace Research, vol.50, no.4, 523-532
\item Högbladh Stina, 2019, “UCDP GED Codebook version 19.1”, Department of Peace and Conflict Research, Uppsala University
}
}
\keyword{datasets}
