% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_clear_df.R
\docType{data}
\name{v_clear_df}
\alias{v_clear_df}
\title{Spaniel and Smith's 𝜈-CLEAR Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8086 rows and 20 columns.}
\source{
\url{https://williamspaniel.com/papers/nuclear/}
}
\usage{
v_clear_df
}
\description{
Abstract: The causes and consequences of nuclear proficiency are central to important questions in international relations. At present, researchers tend to use observable characteristics as a proxy. However, aggregation is a problem: existing measures implicitly assume that each indicator is equally informative and that measurement error is not a concern. We overcome these issues by applying a statistical measurement model to directly estimate nuclear proficiency from observed indicators. The resulting estimates form a new dataset on nuclear proficiency which we call 𝜈-CLEAR. We demonstrate that these estimates are consistent with known patterns of nuclear proficiency while also uncovering more nuance than existing measures. Additionally, we demonstrate how scholars can use these estimates to account for measurement error by revisiting existing results with our measure.
}
\examples{
v_clear_df
dplyr::glimpse(v_clear_df)

}
\references{
\itemize{
\item Smith, Bradley & Spaniel, William. (2018). Introducing ν-CLEAR: a latent variable approach to measuring nuclear proficiency. Conflict Management and Peace Science. 073889421774161. 10.1177/0738894217741619.
}
}
\keyword{datasets}
